yep, chunks finally render, now its time to improve the chunk system overall

+++ ServerCode/src/Chunks/ChunkProvider.cpp
@@ -7,18 +7,18 @@ ChunkProvider::ChunkProvider(ChunkManager* manager) : owningManager(manager) {}
 ChunkProvider::~ChunkProvider() {}
 
 
-Chunk* ChunkProvider::ProvideChunk(int ChunkX, int ChunkY, int ChunkZ, int LOD) {
+Chunk* ChunkProvider::ProvideChunk(int64_t ChunkX, int64_t ChunkY, int64_t ChunkZ, int LOD) {
     {
         std::lock_guard<std::mutex> lock(MTX[LOD]);
         if(IsValidChunk(ChunkX, ChunkY, ChunkZ, LOD)) {
-            return chunks[LOD][glm::ivec3(ChunkX, ChunkY, ChunkZ)];
+            return chunks[LOD][glm::i64vec3(ChunkX, ChunkY, ChunkZ)];
         }
     }
 
     Chunk* c = LoadNewChunk(ChunkX, ChunkY, ChunkZ, LOD);
     {
         std::lock_guard<std::mutex> lock(MTX[LOD]);
-        chunks[LOD][glm::ivec3(ChunkX, ChunkY, ChunkZ)] = c;
+        chunks[LOD][glm::i64vec3(ChunkX, ChunkY, ChunkZ)] = c;
     }
 
     owningManager->GetChunkGenerator().GenerateChunk(c);
@@ -32,11 +32,11 @@ void ChunkProvider::RemoveChunk(Chunk* c) {
     auto& map = GetAllChunks(c->LOD);
     {
         std::lock_guard<std::mutex> lock(MTX[c->LOD]);
-        map.erase(glm::ivec3(c->ChunkX, c->ChunkY, c->ChunkZ));
+        map.erase(glm::i64vec3(c->ChunkX, c->ChunkY, c->ChunkZ));
     }
     delete c;
 }
-std::unordered_map<glm::ivec3, Chunk*>& ChunkProvider::GetAllChunks(int LOD) {
+std::unordered_map<glm::i64vec3, Chunk*>& ChunkProvider::GetAllChunks(int LOD) {
     return chunks[LOD];
 }
 
@@ -44,10 +44,10 @@ std::unordered_map<glm::ivec3, Chunk*>& ChunkProvider::GetAllChunks(int LOD) {
 
 
 
-bool ChunkProvider::IsValidChunk(int ChunkX, int ChunkY, int ChunkZ, int LOD) {
-    return chunks[LOD].find(glm::ivec3(ChunkX, ChunkY, ChunkZ)) != chunks[LOD].end();
+bool ChunkProvider::IsValidChunk(int64_t ChunkX, int64_t ChunkY, int64_t ChunkZ, int LOD) {
+    return chunks[LOD].find(glm::i64vec3(ChunkX, ChunkY, ChunkZ)) != chunks[LOD].end();
 }
-Chunk* ChunkProvider::LoadNewChunk(int ChunkX, int ChunkY, int ChunkZ, int LOD) {
+Chunk* ChunkProvider::LoadNewChunk(int64_t ChunkX, int64_t ChunkY, int64_t ChunkZ, int LOD) {
     Chunk* c = new Chunk;
     c->ChunkX = ChunkX;
     c->ChunkY = ChunkY;
