+++ ServerCode/src/Chunks/ChunkProvider.h
@@ -7,11 +7,11 @@
 
 namespace std {
 	template<>
-	struct hash<glm::ivec3> {
-        std::size_t operator()(const glm::ivec3& k) const noexcept {
-            std::size_t h1 = std::hash<int>()(k.x);
-            std::size_t h2 = std::hash<int>()(k.y);
-            std::size_t h3 = std::hash<int>()(k.z);
+	struct hash<glm::i64vec3> {
+        std::size_t operator()(const glm::i64vec3& k) const noexcept {
+            std::size_t h1 = std::hash<int64_t>()(k.x);
+            std::size_t h2 = std::hash<int64_t>()(k.y);
+            std::size_t h3 = std::hash<int64_t>()(k.z);
 
             // Combine the hashes (boost-style)
             std::size_t seed = h1;
@@ -32,13 +32,13 @@ public:
 
     ChunkManager* owningManager;
 
-    Chunk* ProvideChunk(int ChunkX, int ChunkY, int ChunkZ, int LOD);
+    Chunk* ProvideChunk(int64_t ChunkX, int64_t ChunkY, int64_t ChunkZ, int LOD);
     void RemoveChunk(Chunk* c);
-    std::unordered_map<glm::ivec3, Chunk*>& GetAllChunks(int LOD);
+    std::unordered_map<glm::i64vec3, Chunk*>& GetAllChunks(int LOD);
 private:
-    std::unordered_map<glm::ivec3, Chunk*> chunks[6];
+    std::unordered_map<glm::i64vec3, Chunk*> chunks[6];
     std::mutex MTX[6];
 
-    bool IsValidChunk(int ChunkX, int ChunkY, int ChunkZ, int LOD);
-    Chunk* LoadNewChunk(int ChunkX, int ChunkY, int ChunkZ, int LOD);
+    bool IsValidChunk(int64_t ChunkX, int64_t ChunkY, int64_t ChunkZ, int LOD);
+    Chunk* LoadNewChunk(int64_t ChunkX, int64_t ChunkY, int64_t ChunkZ, int LOD);
 };
